refactor: 서비스에 있던 쿼리 dsl 전용 레포로 분리
+ 쿼리DSL을 위한 전용 레포 추가
+ 응답에 size 누락되던 문제 수정
+ swagger 용 응답 인터페이스 추가
+ backup service의 검색 로직에서 created, updated사용되던 문제 수정
